requirements.txt
pandas
python-dateutil
python-dotenv
requests
